requirements.txt
python-telegram-bot
Flask
Jinja2
Werkzeug
certifi
chardet
gunicorn
requests
pyTelegramBotAPI~=4.2.0
yfinance~=0.1.66
pandas~=1.3.4
pip~=21.1.2
wheel~=0.36.2
lxml~=4.6.4
loguru~=0.5.3
psycopg~=3.0.4
cachetools~=4.2.4
schedule~=1.1.0
seaborn~=0.11.2
Pillow~=8.4.0
